Put the line injection methods into their own separate class
