+++ .pytest_cache/v/cache/nodeids
@@ -73,6 +73,15 @@
   "tests/test_procedure_generation.py::test_full_plate_used[random_test_art_ids2-ccl_artbot_canvas]",
   "tests/test_procedure_generation.py::test_full_plate_used[random_test_art_ids2-ccl_artbot_canvas_90mm_round]",
   "tests/test_procedure_generation.py::test_full_plate_used[random_test_art_ids2-ccl_artbot_canvas_60mm_round]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids0-ccl_artbot_canvas-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids0-ccl_artbot_canvas_90mm_round-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids0-ccl_artbot_canvas_60mm_round-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids1-ccl_artbot_canvas-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids1-ccl_artbot_canvas_90mm_round-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids1-ccl_artbot_canvas_60mm_round-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids2-ccl_artbot_canvas-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids2-ccl_artbot_canvas_90mm_round-1]",
+  "tests/test_procedure_generation.py::test_generate_procedure_8_to_1[random_test_art_ids2-ccl_artbot_canvas_60mm_round-1]",
   "tests/test_user_api.py::test_create_artpiece_by_new_user",
   "tests/test_user_api.py::test_add_user_password[testpass]",
   "tests/test_user_api.py::test_add_user_password[test password]",
